requirements.txt
numpy
pandas
scikit-learn
tensorflow
prefect
python-dotenv
matplotlib
seaborn
flask
joblib